+++ kaparoo/filesystem/search/__init__.py
@@ -1,5 +1,13 @@
 from __future__ import annotations
 
-__all__ = ("search_dirs", "search_files", "search_paths")
+__all__ = (
+    "Selector",
+    "resolve_selector",
+    "search_dirs",
+    "search_files",
+    "search_paths",
+    "select",
+)
 
+from kaparoo.filesystem.search.selection import Selector, resolve_selector, select
 from kaparoo.filesystem.search.wrappers import search_dirs, search_files, search_paths